% Generated by roxygen2 (4.0.2): do not edit by hand
\name{p_dependencies}
\alias{p_dependencies}
\title{Package Dependencies}
\usage{
p_dependencies(package)
}
\arguments{
\item{package}{Character vector containing packages to load.}
}
\value{
Returns a list of package dependencies.
}
\description{
List of package dependencies.
}
\examples{
\dontrun{
pacman::p_load(pacman)
(x<- p_dependencies(pacman))
plot(x)

pacman::p_load(qdap)
(m <- p_dependencies(qdap))
plot(m)
}
}

