% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.p_dependencies}
\alias{plot.p_dependencies}
\title{Plots a p_dependencies Object}
\usage{
\method{plot}{p_dependencies}(x, legend = TRUE, legend.x = -1.5,
  legend.y = -1.05, legend.cex = 0.8,
  title = paste("Dependencies for the", attributes(x)[["package"]],
  "Package"), ...)
}
\arguments{
\item{x}{The p_dependencies object.}

\item{legend}{logical.  If \code{TRUE} a legend is plotted corresponding to
the dependency types.}

\item{legend.x}{the x co-ordinate to be used to position the legend. Can be
specified by keyword or in any way which is accepted by
\code{\link[grDevices]{xy.coords}}.}

\item{legend.y}{the y co-ordinate to be used to position the legend. Can be
specified by keyword or in any way which is accepted by
\code{\link[grDevices]{xy.coords}}.}

\item{legend.cex}{Character expansion factor relative to current
\code{par("cex")}.}

\item{title}{The title of the plot.  Use \code{NULL} to not include a title.}

\item{\ldots}{Arguments passed to \code{\link[graphics]{legend}}.}
}
\description{
Plots a p_dependencies object.
}
\references{
Adapted from mran's Dependencies Graphs \url{http://mran.revolutionanalytics.com/}
}

