% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.p_dependencies}
\alias{print.p_dependencies}
\title{Prints an p_dependencies Object}
\usage{
\method{print}{p_dependencies}(x, ...)
}
\arguments{
\item{x}{The p_dependencies object.}

\item{\ldots}{ignored}
}
\description{
Prints an p_dependencies object.
}

